// client/src/types/posts.d.ts
type IsIntersectingType = 'INTERSECTING' | 'NOT_INTERSECTING'
// ...
type PromptLiterals = 'Show' | 'Discard' | 'Retain' | 'Dommant'

type Theme = 'light' | 'dark';
type ChatOption = 'Hide' | 'Open';
type FontStyle = string | 'font_style';
type ImageTypeProp = 'DP' | 'COVER' | 'NIL'
type NameType = 'photo' | 'coverPhoto' | null
type TargetImageType = {
  name: NameType, 
  data: File | null
}

type CommentOptionProp = { 
  option: ChatOption, 
  storyId: string
}

type EnlargeCompo = {
  type: 'open' | 'enlarge',
  assert: boolean
}

type ConflictType = {
  codeId: string,
  present: boolean
}

type EditingProp = {
  editing: boolean,
  codeId: string,
  code?: string,
  type?: 'DELETE' | 'EDIT' | 'NIL'
}

type UpdateSuccess={
  codeId: string,
  res: boolean
}

type GetConvoType = {
  _id: string, 
  email: string,
  status: Status,
  lastName: string,
  lastSeen: string,
  firstName: string, 
  displayPicture: string,
  userId: string,
  adminId: string,
  isOpened: boolean,
  members: string[],
  membersOpen?: {
    adminOpened: boolean,
    clientOpened: boolean
  }
  createdAt: string,
  updatedAt: string,
  lastMessage: LastMessageType,
}

type InitConversationType = { 
  isLoading: boolean, 
  isError: boolean, 
  msg: string, 
  error: ErrorResponse
}

type LoginPromptType = { opened: ChatOption, source?: 'BadToken' | 'Others' }

type ThemeContextType = {
  theme: Theme,
  parseId: string,
  rollout: boolean,
  fontOption: boolean,
  codeEditor: boolean,
  editing: EditingProp,
  openChat: ChatOption,
  enlarge: EnlargeCompo,
  fontFamily: FontStyle,
  toggleLeft: ChatOption,
  success: UpdateSuccess,
  isPresent: ConflictType,
  currentChat: GetConvoType, 
  openEditPage: ChatOption,
  openNotification: ChatOption,
  loginPrompt: LoginPromptType,
  revealEditModal: ImageTypeProp, 
  notintersecting: IsIntersectingType,
  isConversationState: InitConversationType,
  setTheme: React.Dispatch<React.SetStateAction<Theme>>,
  setParseId: React.Dispatch<React.SetStateAction<string>>,
  setRollout: React.Dispatch<React.SetStateAction<boolean>>,
  setFontFamily: React.Dispatch<React.SetStateAction<string>>,
  setFontOption: React.Dispatch<React.SetStateAction<boolean>>,
  setCodeEditor: React.Dispatch<React.SetStateAction<boolean>>,
  setEditing: React.Dispatch<React.SetStateAction<EditingProp>>,
  setOpenChat: React.Dispatch<React.SetStateAction<ChatOption>>,
  setEnlarge: React.Dispatch<React.SetStateAction<EnlargeCompo>>,
  setToggleLeft: React.Dispatch<React.SetStateAction<ChatOption>>,
  setSuccess: React.Dispatch<React.SetStateAction<UpdateSuccess>>,
  setIsPresent: React.Dispatch<React.SetStateAction<ConflictType>>,
  setOpenEditPage: React.Dispatch<React.SetStateAction<ChatOption>>,
  setCurrentChat: React.Dispatch<React.SetStateAction<GetConvoType>>,
  setOpenNotification: React.Dispatch<React.SetStateAction<ChatOption>>,
  setLoginPrompt: React.Dispatch<React.SetStateAction<LoginPromptType>>,
  setRevealEditModal: React.Dispatch<React.SetStateAction<ImageTypeProp>>,
  setNotIntersecting: React.Dispatch<React.SetStateAction<IsIntersectingType>>,
  setIsConversationState: React.Dispatch<React.SetStateAction<InitConversationType>>
}